NFA: Return empty set when no transition for a state
When we had no transitions for a state in the transition function we
were using null to denote that. This is bad as it causes the library
user to be null aware. A much better option is to return an empty set as
it is also the mathematically correct way to handle it.